# Conflicts:
#	.gitignore
#	becks/projects/admin-app/.gitignore

+++ drupal/API/ab_inbev_api/src/Plugin/rest/resource/UserCustomListResource.php
@@ -2,7 +2,6 @@
 
 namespace Drupal\ab_inbev_api\Plugin\rest\resource;
 
-use Drupal\Core\Session\AccountInterface;
 use Drupal\Component\Plugin\DependentPluginInterface;
 use Drupal\Core\Database\Connection;
 use Drupal\Core\Routing\BcRoute;
@@ -67,13 +66,6 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
   protected $currentRequest;
 
   /**
-   * The current user.
-   *
-   * @var Drupal\Core\Session\AccountInterface;
-   */
-  protected $currentUser;
-
-  /**
    * Vars for the get list
    *
    */
@@ -106,13 +98,11 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
                                 LoggerInterface $logger,
                                 Connection $db_connection,
                                 EntityTypeManagerInterface $entity_type_manager,
-                                Request $request,
-                                AccountInterface $account) {
+                                Request $request) {
     parent::__construct($configuration, $plugin_id, $plugin_definition, $serializer_formats, $logger);
     $this->dbConnection = $db_connection;
     $this->entityTypeManager = $entity_type_manager;
     $this->currentRequest = $request;
-    $this->currentUser = $account;
   }
 
   /**
@@ -127,8 +117,7 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
       $container->get('logger.factory')->get('rest'),
       $container->get('database'),
       $container->get('entity_type.manager'),
-      $container->get('request_stack')->getCurrentRequest(),
-      $container->get('current_user')
+      $container->get('request_stack')->getCurrentRequest()
     );
   }
 
@@ -185,7 +174,7 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
     //   $this->currentUser->isAuthenticated(),
     //   $this->currentUser->getLastAccessedTime()
     // ];
-    // return new ResourceResponse($resp);
+    return new ResourceResponse("");
   }
 
   /**
@@ -200,7 +189,16 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
    *   The HTTP response object.
    */
   public function patch($id, $data) {
+    // $response_array = [];
+
     $uids = json_decode( $data->getContent() , false );
+    $uids = $this->validateUids( $uids );
+
+    if ( empty( $uids ) ) {
+      echo ( "Petición no valida" );
+      return new ModifiedResourceResponse(NULL, 400);
+    }
+
     $patchUids = $this->patchWaitingListUsers( $uids );
     return new ResourceResponse($patchUids);
   }
@@ -217,16 +215,35 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
    * @throws \Symfony\Component\HttpKernel\Exception\HttpException
    */
   public function delete($id, $data) {
+    $uid = $this->currentRequest->get('id');
+
+    // $uids = json_decode( $data->getContent() , false );
+    // $uids = $this->validateUids( $uids );
+    
+    if ( empty( $uid ) ) {
+      echo ( "Petición no valida" );
+      return new ModifiedResourceResponse(NULL, 400);
+    }
+    $retUsers = array();
+    
     $user_storage = $this->entityTypeManager->getStorage('user');
-    $uids = json_decode( $data->getContent() , false );
-    //$retUsers = array();
-    $users = $user_storage->loadMultiple( $uids , TRUE );
-    foreach ($users as $user) {
-      // array_push( $retUsers , $user->get('uid')->value );
+    $user = $user_storage->load( $uid , TRUE );
+    if ( in_array("web_app", $user -> getRoles()) ) {
+      array_push( $retUsers , $user->get('uid')->value );
       $user->delete();
     }
 
+    // $users = $user_storage->loadMultiple( $uids , TRUE );
+    // foreach ($users as $user) {
+    //   if ( in_array("web_app", $user -> getRoles()) ) {
+    //     array_push( $retUsers , $user->get('uid')->value );
+    //     $user->delete();
+    //   }
+    // }
+
     // Deleted responses have an empty body.
+    // echo Response
+    echo( json_encode($retUsers) );
     return new ModifiedResourceResponse(NULL, 202);
   }
 
@@ -301,6 +318,15 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
     // @DCG Add more validation rules here.
   }
 
+  private function validateUids( $uids ) {
+    foreach ($uids as $elementKey => $uid) {
+      if ( $uid == '' || !is_numeric($uid) ) {
+        unset($uids[$elementKey]);
+      }
+    }
+    return $uids;
+  }
+
   /**
    * Loads record from database.
    *
@@ -311,7 +337,7 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
    * @throws \Symfony\Component\HttpKernel\Exception\NotFoundHttpException
    */
   protected function loadRecords() {
-    $user_storage = $this->entityTypeManager->getStorage('user');
+    $user_storage = \Drupal::service('entity_type.manager')->getStorage('user');
 
     $query = $user_storage->getQuery()
               ->condition('roles', 'web_app');
@@ -396,23 +422,19 @@ class UserCustomListResource extends ResourceBase implements DependentPluginInte
     $retUsers = array();
     $users = User::loadMultiple( $uids , TRUE );
     foreach ($users as $user) {
-      if ( $user->get('field_status_waiting_list')->value ) {
+      if ( in_array("web_app", $user -> getRoles()) &&
+            $user->get('field_status_waiting_list')->value ) {
         array_push( $retUsers , $user->get('uid')->value );
         $pass = Util::getRandomUserPass();
         $user->setPassword($pass);
         $user->activate();
         $user->set('field_status_waiting_list', FALSE);
+        $user->set('field_status', 0);  // 0 = normal, 1 = require password change
         if ( $user->save() ) {
-          $mailManager = \Drupal::service('plugin.manager.mail');
-          $module = 'ab_inbev_api';
-          $key = 'default';
-          $to = $user->getEmail();
-          $params['from'] = "contact@becks.com";
-          $params['subject'] = "Bienvenido a Beck's";
-          $params['message'] = "Usuario: ". $to ." <br/> Contraseña: " . $pass ;
-          $langcode = \Drupal::currentUser()->getPreferredLangcode();
-          $send = TRUE;
-          $result = $mailManager->mail($module, $key, $to, $langcode, $params, NULL, $send);
+          Util::sendEmail(  0, 
+                            $user->getEmail() , 
+                            $user->get('field_first_name')->value . ' ' . $user->get('field_last_name')->value
+                          );
         }
       }
     }
